+++ CHANGELOG.md
@@ -13,6 +13,8 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 - RIG-131: Add Search API and database index.
 - RIG-106: Added custom migrations from static websites.
 - RIG-155: Add the Audio media item.
+- RIG-131: Add custom eCMS search form.
+- RIG-131: Theme search page.
 
 ### Changed